backend/pyproject.toml
[project]
name = "attack-path-api"
version = "0.1.0"
description = "FastAPI backend for the attack path visualizer"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0.2",
    "pydantic>=2.9.0",
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
